+++ frb_example/pure_dart/lib/src/rust/api/pseudo_manual/rust_auto_opaque_twin_sync_sse_moi.dart
@@ -264,7 +264,7 @@ class BoxHelloTraitTwinSyncSseMoi extends RustOpaque {
   BoxHelloTraitTwinSyncSseMoi.dcoDecode(List<dynamic> wire)
       : super.dcoDecode(wire, _kStaticData);
 
-  BoxHelloTraitTwinSyncSseMoi.sseDecode(int ptr, int externalSizeOnNative)
+  BoxHelloTraitTwinSyncSseMoi.sseDecode(BigInt ptr, int externalSizeOnNative)
       : super.sseDecode(ptr, externalSizeOnNative, _kStaticData);
 
   static final _kStaticData = RustArcStaticData(
@@ -283,7 +283,7 @@ class BoxMyTraitTwinSyncSseMoi extends RustOpaque {
   BoxMyTraitTwinSyncSseMoi.dcoDecode(List<dynamic> wire)
       : super.dcoDecode(wire, _kStaticData);
 
-  BoxMyTraitTwinSyncSseMoi.sseDecode(int ptr, int externalSizeOnNative)
+  BoxMyTraitTwinSyncSseMoi.sseDecode(BigInt ptr, int externalSizeOnNative)
       : super.sseDecode(ptr, externalSizeOnNative, _kStaticData);
 
   static final _kStaticData = RustArcStaticData(
@@ -303,7 +303,7 @@ class EnumWithGoodAndOpaqueWithoutOptionTwinSyncSseMoi extends RustOpaque {
       : super.dcoDecode(wire, _kStaticData);
 
   EnumWithGoodAndOpaqueWithoutOptionTwinSyncSseMoi.sseDecode(
-      int ptr, int externalSizeOnNative)
+      BigInt ptr, int externalSizeOnNative)
       : super.sseDecode(ptr, externalSizeOnNative, _kStaticData);
 
   static final _kStaticData = RustArcStaticData(
@@ -322,7 +322,8 @@ class NonCloneSimpleEnumTwinSyncSseMoi extends RustOpaque {
   NonCloneSimpleEnumTwinSyncSseMoi.dcoDecode(List<dynamic> wire)
       : super.dcoDecode(wire, _kStaticData);
 
-  NonCloneSimpleEnumTwinSyncSseMoi.sseDecode(int ptr, int externalSizeOnNative)
+  NonCloneSimpleEnumTwinSyncSseMoi.sseDecode(
+      BigInt ptr, int externalSizeOnNative)
       : super.sseDecode(ptr, externalSizeOnNative, _kStaticData);
 
   static final _kStaticData = RustArcStaticData(
@@ -341,7 +342,7 @@ class NonCloneSimpleTwinSyncSseMoi extends RustOpaque {
   NonCloneSimpleTwinSyncSseMoi.dcoDecode(List<dynamic> wire)
       : super.dcoDecode(wire, _kStaticData);
 
-  NonCloneSimpleTwinSyncSseMoi.sseDecode(int ptr, int externalSizeOnNative)
+  NonCloneSimpleTwinSyncSseMoi.sseDecode(BigInt ptr, int externalSizeOnNative)
       : super.sseDecode(ptr, externalSizeOnNative, _kStaticData);
 
   static final _kStaticData = RustArcStaticData(
@@ -430,7 +431,7 @@ class OpaqueOneTwinSyncSseMoi extends RustOpaque {
   OpaqueOneTwinSyncSseMoi.dcoDecode(List<dynamic> wire)
       : super.dcoDecode(wire, _kStaticData);
 
-  OpaqueOneTwinSyncSseMoi.sseDecode(int ptr, int externalSizeOnNative)
+  OpaqueOneTwinSyncSseMoi.sseDecode(BigInt ptr, int externalSizeOnNative)
       : super.sseDecode(ptr, externalSizeOnNative, _kStaticData);
 
   static final _kStaticData = RustArcStaticData(
@@ -449,7 +450,7 @@ class OpaqueTwoTwinSyncSseMoi extends RustOpaque {
   OpaqueTwoTwinSyncSseMoi.dcoDecode(List<dynamic> wire)
       : super.dcoDecode(wire, _kStaticData);
 
-  OpaqueTwoTwinSyncSseMoi.sseDecode(int ptr, int externalSizeOnNative)
+  OpaqueTwoTwinSyncSseMoi.sseDecode(BigInt ptr, int externalSizeOnNative)
       : super.sseDecode(ptr, externalSizeOnNative, _kStaticData);
 
   static final _kStaticData = RustArcStaticData(
@@ -471,7 +472,7 @@ class StructWithGoodAndOpaqueFieldWithoutOptionTwinSyncSseMoi
       : super.dcoDecode(wire, _kStaticData);
 
   StructWithGoodAndOpaqueFieldWithoutOptionTwinSyncSseMoi.sseDecode(
-      int ptr, int externalSizeOnNative)
+      BigInt ptr, int externalSizeOnNative)
       : super.sseDecode(ptr, externalSizeOnNative, _kStaticData);
 
   static final _kStaticData = RustArcStaticData(
